chore: move crypto types to op_crates/crypto (#9609)
